refactor(analysis): split maintainability index owner
