+++ docs/architecture-consolidation-plan.md
@@ -76,6 +76,7 @@ fixtures:
 | Python binding boundary | `crates/tokmd-python/src/lib.rs`, `crates/tokmd-python/src/runtime.rs`, `crates/tokmd-python/src/args.rs`, `crates/tokmd-python/src/envelope.rs`, and `crates/tokmd-python/src/tests.rs` | `lib.rs` 445; runtime owner 175; argument builder owner 59; envelope helper owner 19; binding proof owner 806 | Keep PyO3 wrapper functions and module registration in `lib.rs`; keep shared Python execution and JSON-module handling in `runtime.rs`; keep Python argument-dict construction in `args.rs`; keep envelope extraction and exception mapping in `envelope.rs`; keep Python binding contract tests in `tests.rs`. Preserve the `tokmd_python_binding` proof scope. |
 | Analysis complexity | `crates/tokmd-analysis/src/complexity/mod.rs` + `complexity/functions.rs` + `complexity/functions/` + `complexity/details.rs` + `complexity/details/` + `complexity/summary.rs` + `complexity/risk.rs` + `complexity/debt.rs` + `complexity/histogram.rs` + `complexity/language.rs` + `complexity/math.rs` + `complexity/tests/unit.rs` | 156 + 223 + Rust function owner 126 + 122 + Rust span owner 37 + Python span owner 79 + JS/TS span owner 57 + Go span owner 47 + C-family span owner 52 + 138 + 78 + 69 + 33 + 35 + 5 + 346 | Keep shared complexity logic in `tokmd-analysis`, with Rust function counting plus Rust, Python, JavaScript/TypeScript, Go, and C-family function-span detection in owner modules; split remaining language/source/summary helpers and local unit tests only when fresh evidence shows a clear next seam |
 | Analysis Halstead | `crates/tokmd-analysis/src/halstead/mod.rs`, `halstead/operators.rs`, `halstead/operators/sets.rs`, and `halstead/tokenizer.rs` | report coordinator 205; operator dispatch owner 34; operator set owner 310; tokenizer owner 116 | Keep aggregate report construction in `mod.rs`, language/operator lookup in `operators.rs`, static per-language operator sets in `operators/sets.rs`, and per-file token counting in `tokenizer.rs`. Preserve the internal `crate::halstead::*` test surface and keep changes under the `analysis_halstead` proof scope. |
+| Analysis maintainability | `crates/tokmd-analysis/src/maintainability/mod.rs`, `maintainability/index.rs`, and `maintainability/tests/` | Halstead attachment coordinator 129; index formula owner 65; behavior/proof tests in dedicated test modules | Keep maintainability index formula, grading, rounding, and formula tests in `index.rs`; keep Halstead metric attachment in `mod.rs`; preserve the internal `crate::maintainability::*` test surface and keep changes under the `analysis_maintainability` proof scope. |
 | Analysis near-duplicate detection | `crates/tokmd-analysis/src/near_dup/mod.rs`, `near_dup/clusters.rs`, `near_dup/fingerprint.rs`, and `near_dup/pairs.rs` | report coordinator 274; cluster owner 305; fingerprint owner 95; pair scoring owner 186 | Keep report assembly, eligibility filtering, fingerprint orchestration, and scope partitioning in `mod.rs`; keep cluster construction in `clusters.rs`; keep Winnowing tokenization, k-gram hashing, file reads, and fingerprint selection in `fingerprint.rs`; keep inverted indexing, shared-fingerprint counts, Jaccard scoring, and deterministic pair ordering in `pairs.rs`. Preserve the `analysis_near_dup` proof scope. |
 | CLI parser | `crates/tokmd/src/cli/parser.rs` and `crates/tokmd/src/cli/parser/` | `parser.rs` 209; command enum owner 103; global scan args owner 133; shared value-enum owner 235; analyze parser owner 178; context/handoff parser owner 166; cockpit/baseline parser owner 89; diff parser owner 67; gate parser owner 55; sensor parser owner 43; export parser owner 43; badge parser owner 45; module parser owner 36; lang parser owner 26; run parser owner 25; check-ignore parser owner 15; completions parser owner 21; init parser owner 36; tools parser owner 15 | Context, handoff, analyze, cockpit, baseline, diff, gate, sensor, export, badge, module, lang, run, check-ignore, completions, init, tools, command enum, global scan args, and shared value enums now live under parser owner modules; continue command-family splits only when clap snapshots prove behavior is unchanged |
 | CLI config resolution | `crates/tokmd/src/config.rs`, `crates/tokmd/src/config/resolve.rs`, and `crates/tokmd/src/config/resolve/` | `config.rs` 313; resolver barrel 10; lang owner 123; module owner 148; export owner 184; parse owner 35 | Keep config discovery, profile selection, `ConfigContext`, and `ResolvedConfig` in the root config module while lang/module/export CLI-to-receipt argument resolution and string-to-enum parsing live in focused resolver owner modules. Preserve root re-exports. |
